Jobs_Scheduling/scheduler_microservice/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-cron==0.5.1
croniter==1.0.15
drf-yasg==1.20.0